c/scheduling: moved allocation_node to be top level type
